refactor tag queryKey config로 분리
