+++ src/features/search/config/queryKeys.ts
@@ -1,0 +1,3 @@
+export const QUERY_KEYS = {
+  GET_TAG: () => ["tags"],
+} as const
